Allows trailing whitespace before closing > in inline/block html
Previously the regexp expected the closing first tag to contain zero
whitespace between the last property and the closing angle bracket.

As a result, the code `<a href="bar" >foo</a>` would render incorrectly
as `<p><a href="bar" >foo&lt;/a&gt;</p>\n`.

This change applies to unescaped inline and block elements.

Fixes #104