[fix] Add runtime error if renderer is not an instance of BaseRenderer
I had a case where I was passing the class itself, and not an instance of the
class.  It took a bit to figure out what the issue was since the error was
about not passing data to renderer.finalize() -- when it clearly was passing it.

I thought this might make the project a bit more robust by adding this check.