Use absolute import in Spoiler example, add markdown syntax
Examples are more helpful if they contain absolute imports, so I can copy/paste them in my own project.

An example of how to use the spoiler tag is also nice to understand the mechanism. This is available in the general sense in https://mistune.readthedocs.io/en/latest/directives.html, but an explicit example here would make it more understandable.